src/main/java/com/gamesys/registration/validation/UsernameValidator.java
package com.gamesys.registration.validation;

import org.springframework.util.StringUtils;

import javax.validation.ConstraintValidator;
import javax.validation.ConstraintValidatorContext;

public class UsernameValidator implements ConstraintValidator<Username, String> {

    @Override
    public boolean isValid(String username, ConstraintValidatorContext constraintValidatorContext) {
        return StringUtils.hasLength(username) && username.chars().allMatch(Character::isLetterOrDigit);
    }
}
